Adding exogenous_*_predictability parameters
Using these parameters SCML2020/2021/OneShot worlds can be configured so
that the exogenous contracts of the same agents at different time-steps
are predictable (or not). If predictability is zero then each agent can
have any quantity for its exogenous contracts. If predictability is 1.0
then it will have the same quantity at every time-step and if it is
somewhere in between, the quantity at different steps will be similar to
each other.