Adding before_step() to all agents.
If a method called `before_step()` is defined for an agent, it will be
called once every simulated day before any other calls to the agent but
after all exogenous contracts and ufun parameters are fixed.

Note that `step()` is called at the *end* not the beginning of the day.