Resetting agents before making negotiations
before_step() was also resetting agents in oneshot. This meant that all
negotiators that were created in _make_negotiations() were deleted in
reset().

Now we call  reset() before making negotaitions and before_step() after
that.

This guarantees that the agent has access to its negotiators in
before_step() as in 2020.