Adding wide_price_range to control the unit-price range
When creating negotiation issues, the trading price of the previous
product will be used for seting the minimum price in the negotiation
agenda if wide_price_range is pased (default) and for the same product
if it is not passed.

Passing wide_price_ragne=False when constructing SCML2020OneShotWorld
will increase the importance of quantity and reduce the importance of
price.