Adding RewardFunction to OneShotEnv
The environment now receives a reward function object (implementing the
RewardFunction protocol defined in oneshot.rl.reward.py) which is called
as follows every step of the environment simulation:

1. Before the agent is asked to act, the method `before_action` is
   called which can return any kind of information to be used later in
   the reward calculation.
2. After the agent acts, the reward function is called (i.e. __call__ is
   called) with the awi, the action and whatever information was
   generated by `before_action`

The DefaultRewardFunction simply uses the difference in the balance
before and after the action to define the reward (balance after -
balance before). This implies that the reward is sparse and is only
given at the end of all negotiations in a simulated day.